Bump up warnings
- Make Verilator warnings back into errors and bump minimum Verilator
version to address erroneous warning
- Make warnings errors for host tool build
- Remove array style pointer args in emulator

+++ tools/serial_boot/serial_boot.c
@@ -111,7 +111,7 @@ int read_serial_long(int serial_fd, unsigned int *out, int timeout)
 		if (!read_serial_byte(serial_fd, &ch, timeout))
 			return 0;
 
-		result = (result >> 8) | (ch << 24);
+		result = (result >> 8) | ((unsigned int) ch << 24);
 	}
 	
 	*out = result;
